fix: code review round 3 — multinomial dtype, cache limit, EMA semantics
Fixes:
- [M2] sample_pairs_gpu: cast to float32 before torch.multinomial to
  avoid crashes on float64 inputs (GPU multinomial requires float32)
- [M3] DijkstraProvider: cap cache at 2000 rows per side with FIFO
  eviction to prevent unbounded memory growth on large graphs
- [M6] lambda_ema_beta=0.0 now disables EMA (same as None), fixing
  the unintuitive behavior of locking to first iteration's Lambda
- [m1] Remove redundant `not verbose` in _sinkhorn_loop compiled path
- [m5] C_lin_device uses float64 to match Lambda_gw dtype, avoiding
  implicit upcasting every iteration in the FGW blending step
- [T1] Add test for cost plateau early stopping (verifies n_iter < max_iter)

+++ torchgw/_sampling.py
@@ -75,13 +75,13 @@ def sample_pairs_gpu(
         rng = np.random.default_rng()
         return rng.integers(0, N, size=M), rng.integers(0, K, size=M)
 
-    p_rows = p_rows / total
+    p_rows = (p_rows / total).float()  # multinomial requires float32
     rows = torch.multinomial(p_rows, M, replacement=True)
 
     # Column sampling per selected row
     row_probs = T[rows]  # (M, K)
     row_sums = row_probs.sum(dim=1, keepdim=True).clamp(min=1e-30)
-    row_probs = row_probs / row_sums
+    row_probs = (row_probs / row_sums).float()
     cols = torch.multinomial(row_probs, 1).squeeze(1)
 
     return rows.cpu().numpy(), cols.cpu().numpy()
